fix(pointcloud_concatenate): bug in look up function

+++ src/simple_point_concatenate/src/pointcloud_concatenate/offset_concatenate_pointclouds.cpp
@@ -298,11 +298,11 @@ bool PointCloudOffsetConcatenationComponent::all_points_received()
 
 int PointCloudOffsetConcatenationComponent::lookup_index(const std::vector<double> & stamp_buff, const double stamp)
 {
-  constexpr int tolerance_msec = 50;
-  int min_offset_msec = tolerance_msec;
+  constexpr double tolerance_msec = 50.0;
+  double min_offset_msec = tolerance_msec;
   int min_offset_index = -1;
   for (size_t index = 0; index < stamp_buff.size(); ++index) {
-    int offset_msec = std::abs(stamp_buff[index] - stamp);
+    double offset_msec = std::abs(stamp_buff[index] - stamp);
     if (offset_msec <= tolerance_msec && offset_msec < min_offset_msec) {
       min_offset_msec = offset_msec;
       min_offset_index = index;
